requirements.txt
kaggle>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0
numpy
torch
torchvision
xformers
pandas>=2.0
transformers>=4.40